Rearrange the codes to their component file
