+++ Frontend/src/packages/editor/Editor.tsx
@@ -1,166 +1,13 @@
-import React, {ReactNode, useCallback, useMemo, useState} from 'react';
-import { EditorBoldButton, EditorItalicButton, EditorUnderlineButton } from "./components/buttons";
+import React from 'react';
+import ContentBlock from "./components/ContentBlock";
 
-// slate-js dependencies
-import {Editable, Slate, useSlate, withReact} from "slate-react";
-import { createEditor, Descendant, Editor as SlateEditor } from "slate";
-import styled from "styled-components";
-import {withHistory} from "slate-history";
-
-
-type RenderLeafProps = {
-  attributes: any
-  children: ReactNode
-  leaf: {
-    bold?: boolean
-    italic?: boolean
-    underline?: boolean
-  }
-}
-
-const Toolbar = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 20px;
-  align-items: center;
-  justify-content: flex-start;
-  margin: 10px 20px;
-  height: fit-content;
-  outline: 1px solid black;
-`
 
 const Editor = () => {
 
-  const renderElement = useCallback(props => <Element { ...props } />, []);
-  const renderLeaf = useCallback(props => < Leaf { ...props } />, []);
-  const editor = useMemo(() => withHistory(createEditor()), []);
-
   return (
-    <Slate editor={editor} value={initialValue}>
-      <Toolbar>
-        <EditorBoldButton />
-        <EditorItalicButton />
-        <EditorUnderlineButton />
-      </Toolbar>
-      <Editable
-        renderElement={renderElement}
-        renderLeaf={renderLeaf}
-      />
-    </Slate>
+    <ContentBlock />
   );
 };
 
-
-const Element = (props: any) => {
-  const { attributes, children, element } = props
-  const style = { textAlign: element.align }
-  switch (element.type) {
-    case 'block-quote':
-      return (
-        <blockquote style={style} {...attributes}>
-          {children}
-        </blockquote>
-      )
-    case 'bulleted-list':
-      return (
-        <ul style={style} {...attributes}>
-          {children}
-        </ul>
-      )
-    case 'heading-one':
-      return (
-        <h1 style={style} {...attributes}>
-          {children}
-        </h1>
-      )
-    case 'heading-two':
-      return (
-        <h2 style={style} {...attributes}>
-          {children}
-        </h2>
-      )
-    case 'list-item':
-      return (
-        <li style={style} {...attributes}>
-          {children}
-        </li>
-      )
-    case 'numbered-list':
-      return (
-        <ol style={style} {...attributes}>
-          {children}
-        </ol>
-      )
-    default:
-      return (
-        <p style={style} {...attributes}>
-          {children}
-        </p>
-      )
-  }
-};
-
-const Leaf = (props: RenderLeafProps) => {
-  // eslint-disable-next-line prefer-const
-  let { attributes, children, leaf } = props
-  if (leaf.bold === true) {
-    children = <strong>{children}</strong>
-  }
-
-  if (leaf.italic === true) {
-    children = <em>{children}</em>
-  }
-
-  if (leaf.underline === true) {
-    children = <u>{children}</u>
-  }
-
-  return <span {...attributes}>{children}</span>
-};
-
-const isMarkActive = (editor: SlateEditor, format: string): boolean => {
-  const marks = SlateEditor.marks(editor);
-  return marks ? marks[format as keyof typeof marks] === true : false;
-};
-
-const toggleMark = (editor: SlateEditor, format: string) => {
-  const isActive = isMarkActive(editor, format);
-
-  if (isActive) {
-    SlateEditor.removeMark(editor, format)
-  } else {
-    SlateEditor.addMark(editor, format, true)
-  }
-};
-
-const initialValue: Descendant[] = [
-  {
-    type: 'heading-two',
-    children: [
-      { text: 'A heading' },
-    ],
-  },
-  {
-    type: 'paragraph',
-    children: [
-      {
-        text: "This is a paragraph.",
-      },
-      {
-        text: "bold",
-        bold: true
-      }
-    ],
-  },
-  {
-    type: 'block-quote',
-    children: [{ text: 'A block quote.' }],
-  },
-  {
-    type: 'paragraph',
-    align: 'center',
-    children: [{ text: 'text in the center' }],
-  },
-]
 
 export default Editor;
